% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc2.R
\name{analyze}
\alias{analyze}
\alias{analyze.wroc}
\title{Optimal Bucketing With a Fixed Number of Buckets}
\usage{
analyze(x, ...)

\method{analyze}{wroc}(x, nbuckets = 5, trend = c("auto", "upper", "lower"))

analyze(x, ...)

\method{analyze}{wroc}(x, nbuckets = 5, trend = c("auto", "upper", "lower"))
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{nbuckets}{The number of buckets the final bucketing should have.}

\item{trend}{The trend according to the cummulative ROC curve.}

\item{x}{An object of class \code{wroc}.}

\item{nbuckets}{The number of buckets the final bucketing should have.}

\item{trend}{The trend according to the cummulative ROC curve.}
}
\value{
An object of class \code{wroc} but with the original bucket numbers.
  The class \code{analyze.wroc} was added to show the curve has been
  modified. If the input is a \code{wroc.list} then the output is again a
  \code{wroc.list} of constraint-optimized ROC curves.

An object of class \code{wroc} but with the original bucket numbers.
  The class \code{analyze.wroc} was added to show the curve has been
  modified. If the input is a \code{wroc.list} then the output is again a
  \code{wroc.list} of constraint-optimized ROC curves.
}
\description{
Optimizes the bucketing to have the maximum Gini index possible for the given
trend and number of buckets.

Optimizes the bucketing to have the maximum Gini index possible for the given
trend and number of buckets.
}
\section{Methods (by class)}{
\itemize{
\item \code{wroc}: Method for a single variable.

\item \code{wroc}: Method for a single variable.
}}

