% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{analyze.wroc}
\alias{analyze.wroc}
\title{Optimal Bucketing With a Fixed Number of Buckets}
\usage{
\method{analyze}{wroc}(x, nbuckets = 5, trend = c("auto", "upper", "lower"))
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{nbuckets}{The number of buckets the final bucketing should have.}

\item{trend}{The trend according to the cummulative ROC curve.}
}
\description{
Optimizes the bucketing to have the maximum Gini index possible for the given
trend and number of buckets.
}

