% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc_list.R
\name{optimize}
\alias{optimize}
\alias{optimize.wroc}
\alias{optimize.wroc.list}
\title{Optimal Gini Bucketing}
\usage{
optimize(x, ...)

\method{optimize}{wroc}(x, trend = c("auto", "upper", "lower"),
  method = c("magic", "optimal", "cascade"), min_p_pob = 0.05)

\method{optimize}{wroc.list}(x, trends = "auto", method = c("magic",
  "optimal", "cascade"), min_p_pob = 0.05, verbose = TRUE)
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{trend}{The trend according to the cummulative ROC curve.}

\item{method}{One of "optimal" (optimal Gini regardless of population),
"magic" [default] (tries to optimize Gini while keeping the distribution as
homogeneous as possible) and "cascade" (iteratively pastes the smallest
bucket left or right so as to lose as little Gini as possible).}

\item{min_p_pob}{Numeric between 0 and 1. Minimum proportion of population to
keep in each bucket. Only for methods "magic" and "cascade",}

\item{trends}{A character vector containing the trend to choose for each
variable or 'auto' to let the algorithm find out on its own.}
}
\value{
An object of class \code{wroc}. The class \code{optimize.wroc} has
  been added to show the fact that the original curve has been optimized. If
  the input is a \code{wroc.list} then the output is again a \code{wroc.list}
  of optimized ROC curves.
}
\description{
Optimizes the bucketing to have the maximum Gini index possible for the given
trend.
}
\section{Methods (by class)}{
\itemize{
\item \code{wroc}: Optimize a single ROC curve.

\item \code{wroc.list}: Runs the algorithm on a \code{wroc.list}.
}}
\seealso{
\code{analyze.wroc}
}

