% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{performance}
\alias{performance}
\alias{performance.numeric}
\alias{performance.wroc}
\title{Performance Metrics for ROC Curves}
\usage{
performance(x, ...)

\method{performance}{wroc}(x)

\method{performance}{numeric}(predictions, labels, ...)
}
\arguments{
\item{x}{An object of type \code{wroc}.}

\item{predictions}{A numeric vector or a matrix with one column with the
predictions of a score or the value of a variable whose ROC curve is to be
created.}

\item{labels}{Either a factor (or character vector) with the true labels of
the response variable or a 2-column matrix containing counts for each class
for each distinct value in \code{predictions}.}
}
\description{
Calculates several performance metrics for ROC curves, including AUC and Gini
index for both trends.
}
\section{Methods (by class)}{
\itemize{
\item \code{wroc}: Performance measures of a \code{wroc} object.

\item \code{numeric}: Quick evaluation without need of a \code{wroc}
object.
}}

