% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc_list.R
\name{plot.wroc.list}
\alias{plot.wroc.list}
\title{aaaa}
\usage{
\method{plot}{wroc.list}(x, type = c("accum", "roc", "trend", "woe"),
  include.special = TRUE, save.pdf = TRUE, file = "Wroc Plots.pdf", ...)
}
\arguments{
\item{save.pdf}{Should the plots be saved to a PDF file or just shown?}

\item{file}{Name (with path) of the PDF file to output the plots to}

\item{...}{Additional options to pass on to \code{pdf()}}
}
\description{
sadfsfd
}

