% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc_list.R
\name{plot.wroc}
\alias{plot.wroc}
\alias{plot.wroc.list}
\title{Plot ROC Curves}
\usage{
\method{plot}{wroc}(x, type = c("accum", "roc", "default", "woe"),
  include.special = TRUE, label.size = 3)

\method{plot}{wroc.list}(x, type = c("accum", "roc", "default", "woe"),
  include.special = TRUE, save.pdf = FALSE, file = "Wroc Plots.pdf", ...)
}
\arguments{
\item{x}{An object of class \code{wroc} or \code{wroc.list}.}

\item{type}{One of 'accum' (ROC using fnr ~ tnr), 'roc' (standard ROC tpr ~
fpr), 'default' (probability of being in the positive class) and 'woe'
(Weight of Evidence: log-odds of being in the *negative* class).}

\item{include.special}{Should special values be included in the plot? Only
applies to 'default' and 'woe' options}

\item{save.pdf}{Should the plots be saved to a PDF file or just shown?}

\item{file}{Name (with path) of the PDF file to output the plots to}

\item{...}{Additional options to pass on to \code{pdf()}}
}
\value{
An object of class ggplot which can be then be modified if needed. If
  \code{x} is a \code{wroc.list}, then a list of \code{ggplot} objects is
  returned. Depending on the value of \code{save.pdf}, a PDF with the plots
  is saved to \code{file} or the plots are shown one by one.
}
\description{
Uses \code{ggplot2} to plot several interesting aspects about a ROC curve.
}

