% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{points.wroc.list}
\alias{points.wroc.list}
\title{Transformation to Points}
\usage{
\method{points}{wroc.list}(x, model, base.points = 600, base.odds = 1/20,
  pdo = 15, point.decimals = 0, prefix.to.remove = "woe_")
}
\description{
Transformation to Points
}

