% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{predict.wroc}
\alias{predict.wroc}
\title{Transform Variables with a ROC Curve}
\usage{
\method{predict}{wroc}(object, newdata, type = c("woe", "bucket", "p_bad"))
}
\arguments{
\item{object}{An object of class \code{wroc}.}

\item{newdata}{A numeric vector containing values of the raw variable to be
transformed.}

\item{type}{A string containing the name of the column in \code{object$info},
to be used as transformation, usually 'woe', 'bucket' or 'p_bad'.}
}
\description{
Transforms raw variables into bucketed ones. The pasted value can be any
column in the \code{object$info} \code{data.frame}, but it usually is the
WoE, the bucket number or the probability of being in the positive class.
}

