% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{print.summary.wroc}
\alias{print.summary.wroc}
\title{Important Information of ROC Curves}
\usage{
\method{print}{summary.wroc}(x, extended = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{extended}{Should details be printed?}
}
\description{
Returns and prints basic important information about a \code{wroc} object.
}

