% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc2.R
\name{reset.buckets}
\alias{reset.buckets}
\title{Reset Bucket Numbers of \code{wroc} Objects}
\usage{
reset.buckets(x)

reset.buckets(x)
}
\arguments{
\item{x}{An object of type \code{wroc}.}

\item{x}{An object of type \code{wroc}.}
}
\description{
Resets the bucket numbers of an optimized or subsetted ROC curve in order to
have consecutive bucket numbers.

Resets the bucket numbers of an optimized or subsetted ROC curve in order to
have consecutive bucket numbers.
}

