% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc_list.R, R/wroc_list2.R
\name{subset.wroc.list}
\alias{[.wroc.list}
\alias{subset.wroc.list}
\title{Choose ROC Curves Manually from a List}
\usage{
\method{subset}{wroc.list}(x, keep = NULL, drop = NULL)

\method{[}{wroc.list}(x, i)

\method{subset}{wroc.list}(x, keep = NULL, drop = NULL)

\method{[}{wroc.list}(x, i)
}
\arguments{
\item{x}{An object of type \code{wroc.list}}

\item{keep}{A character vector of the names of the ROC curves to be kept
(overrides \code{drop})}

\item{drop}{A character vector of the names of the ROC curves to be discarded
(is overriden by \code{keep})}

\item{x}{An object of type \code{wroc.list}}

\item{keep}{A character vector of the names of the ROC curves to be kept
(overrides \code{drop})}

\item{drop}{A character vector of the names of the ROC curves to be discarded
(is overriden by \code{keep})}
}
\description{
Provides a 'keep' and 'drop'-style for selecting ROC curves in a large list.

Provides a 'keep' and 'drop'-style for selecting ROC curves in a large list.
}

