% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{subset.wroc}
\alias{subset.wroc}
\title{Paste Adjacent Buckets of a ROC Curve}
\usage{
\method{subset}{wroc}(x, buckets = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{buckets}{Bucket numbers to be pasted to the right.}
}
\value{
An object of class \code{wroc} with the given buckets pasted to the
  right. The class \code{subset.wroc} has been added to show that the ROC
  curve isn't the original one. The original bucket numbers are used for
  clarity (you can use reset.buckets afterwards to change this)
}
\description{
Pastes the given bucket names to the immediately following one. This allows
for manual tampering of the ROC curves.
}

