% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc_list.R
\name{summary.wroc}
\alias{print.summary.wroc}
\alias{print.summary.wroc.list}
\alias{summary.wroc}
\alias{summary.wroc.list}
\title{Important Information About ROC Curves}
\usage{
\method{summary}{wroc}(object, performance = TRUE, ...)

\method{print}{summary.wroc}(object, extended = FALSE, ...)

\method{summary}{wroc.list}(x, performance = T, ...)

\method{print}{summary.wroc.list}(x, ...)
}
\arguments{
\item{object}{An object of class \code{wroc}.}

\item{performance}{Calculate performance measures as given by
\code{performance} (AUC, Gini, KS, IV)?}

\item{extended}{Should details be printed?}
}
\value{
An object of class \code{summary.wroc}, which has interesting
  non-internal information about the ROC curve and prints prettily. If
  performance = FALSE then performance is not run and hence the method will
  be faster for ROC curves with a very large number of buckets.
}
\description{
Returns and prints basic important information about a \code{wroc} or
\code{wroc.list} object.
}

