% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{woe}
\alias{woe}
\title{Calculate Weight of Evidence}
\usage{
woe(n_good, n_bad, na.factor = 2, extremum.woe = 5)
}
\arguments{
\item{n_good}{Number of negative examples ("goods")}

\item{na.factor}{Multiplicative facor in case some cases are NAs}

\item{extremum.woe}{WoE in case all observations are NA}

\item{n_good}{Number of positive examples ("bads")}
}
\description{
Calculates the Weight of Evidence, namely \code{log((n_good/Population)/(n_bad/Population))}, while handling special cases (either no goods, no bads or both). If all the buckets are special, the value defaults to \code{abs(extremum.woe)} if there are only goods, \code{-abs(extremum.woe)} if there are only goods and 0 if there are neither. If only a few buckets have no goods or bads, then the values default to \code{na.factor} times the smallest (if no goods) or highest present WoE(if no bads).
}

