% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{wroc}
\alias{c.wroc}
\alias{plot.wroc}
\alias{print.wroc}
\alias{wroc}
\alias{wroc.default}
\alias{wroc.formula}
\title{Create Flexible ROC Curves}
\usage{
wroc(x, ...)

wroc.default(predictions, labels, ngroups = NULL, level.bad = 1,
  col.bad = 1, special.values = NULL)

wroc.formula(formula, data, ngroups = NULL, level.bad = 1,
  special.values = NULL, verbose = (nrow(data) > 10000))

\method{print}{wroc}(x, ...)

\method{c}{wroc}(...)

\method{plot}{wroc}(x, type = c("accum", "roc", "trend", "woe"),
  include.special = TRUE)
}
\arguments{
\item{x}{An object of class \code{wroc}}

\item{predictions}{A numeric vector or a matrix with one column with the
predictions of a score or the value of a variable whose ROC curve is to be
created.}

\item{labels}{Either a factor (or character vector) with the true labels of
the response variable or a 2-column matrix containing counts for each class
for each distinct value in \code{predictions}.}

\item{ngroups}{Precision to use for the ROC curve. A NULL value means "use as
many values as there are unique values in \code{predictions}".}

\item{level.bad}{The level or value in \code{labels} corresponding to the
positive class.}

\item{col.bad}{If \code{labels} is a matrix, the column containing the counts
for the positive class.}

\item{special.values}{A vector containing values to be treated separately
(NAs should be replaced by one such value).}

\item{formula}{Standard formula to specify the response (on the left side)
and the variables whose ROC curves are to be created.}

\item{data}{A \code{data.frame} with the information.}

\item{type}{One of 'accum' (ROC using fnr ~ tnr), 'roc' (standard ROC tpr ~
fpr), 'trend' (probability of being in the positive class) and 'woe'
(Weight of Evidence: log-odds of being in the *negative* class).}

\item{include.special}{Should special values be included in the plot? Only
applies to 'trend' and 'woe' options}
}
\value{
An object of class \code{wroc} (default method) or a \code{wroc.list}
  of such objects (formula method), which can be then be used for several
  purposes.
}
\description{
\code{wroc} builds a ROC curve based on either a raw variable or a summarized
dataset. \code{wroc} objects can be manipulated in a variety of ways,
including manual and Optimal Gini bucketing.
}
\section{Functions}{
\itemize{
\item \code{wroc.default}: Method for raw, unsummarized data

\item \code{wroc.formula}: Formula-based for simple processing of several variables at
once.
}}

