% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc_list.R
\name{performance}
\alias{performance}
\alias{performance.numeric}
\alias{performance.wroc}
\alias{performance.wroc.list}
\title{Performance Metrics for ROC Curves}
\usage{
performance(x, ...)

\method{performance}{wroc}(x)

\method{performance}{numeric}(predictions, labels, ...)

\method{performance}{wroc.list}(x, out.type = c("list", "data.frame"))
}
\arguments{
\item{x}{An object of type \code{wroc} or \code{wroc.list}.}

\item{...}{Aditional arguments for \code{wroc}}

\item{predictions}{A numeric vector or a matrix with one column with the
predictions of a score or the value of a variable whose ROC curve is to be
created.}

\item{labels}{Either a factor (or character vector) with the true labels of
the response variable or a 2-column matrix containing counts for each class
for each distinct value in \code{predictions}.}
}
\value{
A list with several performance statistics *based on the inverse ROC
  curve* (as given by the 'trend' option in \code{plot.wroc}), that is fnr ~
  tnr:

  \itemize{ \item{\code{auc}: Area under the ROC curve}
  \item{\code{gini_two_way}: abs(2*auc - 1) (different from normal Gini when
  the ROC curves crosses the identity line)} \item{\code{gini_up}: Gini with
  an upper trend} \item{\code{gini_down}: Gini with a lower trend}
  \item{\code{gini}: The larger of gini_up and gini_down}
  \item{\code{best_trend}: The trend of the larger Gini} \item{\code{ks}:
  Kolmogorov-Smirnov statistic} \item{\code{iv}: Information Value} } In the
  case where \code{x} is a \code{wroc.list}, then a list of the
  aforementioned is returned instead.
}
\description{
Calculates several performance metrics for ROC curves, including AUC and Gini
index for both trends.
}
\section{Methods (by class)}{
\itemize{
\item \code{wroc}: Performance measures of a \code{wroc} object.

\item \code{numeric}: Quick evaluation without need of a \code{wroc}
object.

\item \code{wroc.list}: Method for a list of ROC curves
}}

