% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{plot.wroc}
\alias{plot.wroc}
\title{Plot ROC Curves}
\usage{
\method{plot}{wroc}(x, type = c("accum", "roc", "trend", "woe"),
  include.special = TRUE)
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{type}{One of 'accum' (ROC using fnr ~ tnr), 'roc' (standard ROC tpr ~
fpr), 'trend' (probability of being in the positive class) and 'woe'
(Weight of Evidence: log-odds of being in the *negative* class).}

\item{include.special}{Should special values be included in the plot? Only
applies to 'trend' and 'woe' options}
}
\value{
An object of class ggplot which can be then be modified if needed.
}
\description{
Uses \code{ggplot2} to plot several interesting aspects about a ROC curve.
}

