% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{subset.wroc}
\alias{subset.wroc}
\title{Paste Adjacent Buckets of a ROC Curve}
\usage{
\method{subset}{wroc}(x, buckets = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{wroc}.}

\item{buckets}{Bucket numbers to be pasted to the right.}
}
\description{
Pastes the given bucket names to the immediately following one. This allows
for manual tampering of the ROC curves.
}

