% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R
\name{summary.wroc}
\alias{summary.wroc}
\title{Important Information of ROC Curves}
\usage{
\method{summary}{wroc}(object, performance = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{wroc}.}

\item{performance}{Calculate performance measures as given by
\code{performance} (AUC, Gini, KS, IV)?}
}
\description{
Returns and prints basic important information about a \code{wroc} object.
}

