% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wroc.R, R/wroc_list.R
\name{wroc}
\alias{c.wroc}
\alias{c.wroc.list}
\alias{print.wroc}
\alias{print.wroc.list}
\alias{wroc}
\alias{wroc.default}
\alias{wroc.formula}
\title{Create Flexible ROC Curves}
\usage{
wroc(x, ...)

\method{wroc}{default}(predictions, labels, ngroups = NULL, level.bad = 1,
  col.bad = 1, special.values = NULL)

\method{wroc}{formula}(formula, data, ngroups = NULL, level.bad = 1,
  special.values = NULL, verbose = (nrow(data) > 10000))

\method{print}{wroc}(x)

\method{c}{wroc}(...)

\method{print}{wroc.list}(x, ...)

\method{c}{wroc.list}(...)
}
\arguments{
\item{predictions}{A numeric vector or a matrix with one column with the
predictions of a score or the value of a variable whose ROC curve is to be
created.}

\item{labels}{Either a factor (or character vector) with the true labels of
the response variable or a 2-column matrix containing counts for each class
for each distinct value in \code{predictions}.}

\item{ngroups}{Precision to use for the ROC curve. A NULL value means "use as
many values as there are unique values in \code{predictions}".}

\item{level.bad}{The level or value in \code{labels} corresponding to the
positive class.}

\item{col.bad}{If \code{labels} is a matrix, the column containing the counts
for the positive class.}

\item{special.values}{A vector containing values to be treated separately
(NAs should be replaced by one such value).}

\item{formula}{Standard formula to specify the response (on the left side)
and the variables whose ROC curves are to be created.}

\item{data}{A \code{data.frame} with the information.}
}
\value{
An object of class \code{wroc} (default method) or a \code{wroc.list}
  of such objects (formula method and combine method), which can be then be
  manipulated in bulk.
}
\description{
\code{wroc} build ROC curves based on either a raw variable or a summarized
dataset. \code{wroc} objects can be manipulated in a variety of ways,
including manual and Optimal Gini bucketing.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method for raw, unsummarized data

\item \code{formula}: Formula-based for simple processing of several variables at
once.
}}

